+++ src/_P019_PCF8574.ino
@@ -40,7 +40,7 @@ TaskDevicePluginConfigLong settings:
 boolean Plugin_019(byte function, struct EventStruct *event, String& string)
 {
   boolean success = false;
-  static int8_t switchstate[TASKS_MAX];
+  static int switchstate[TASKS_MAX];
 
   switch (function)
   {
@@ -121,7 +121,7 @@ boolean Plugin_019(byte function, struct EventStruct *event, String& string)
     case PLUGIN_INIT:
       {
         // read and store current state to prevent switching at boot time
-        switchstate[event->TaskIndex] = Plugin_019_Read(Settings.TaskDevicePort[event->TaskIndex]);
+        switchstate[event->TaskIndex] = (int) Plugin_019_Read(Settings.TaskDevicePort[event->TaskIndex]);
 
         // @giig1967g-20181022: set initial UserVar of the switch
         if (Settings.TaskDevicePin1Inversed[event->TaskIndex]){
